software/rpi/requirements.txt
paho-mqtt
pyserial
pymavlink
